+++ app/controllers/admin/page_controller.rb
@@ -1,0 +1,6 @@
+module Admin
+  class PageController < BaseAdminController
+    def index
+    end
+  end
+end
