--- app/controllers/admin_controller.rb
@@ -1,12 +1,0 @@
-class AdminController < ApplicationController
-  before_action :ensure_authenticated_user
-
-  def index
-  end
-
-  private
-
-  def ensure_authenticated_user
-    redirect_to sign_in_path unless signed_in?
-  end
-end
