+++ app/views/admin/auth/sign_in.html.erb
@@ -3,6 +3,6 @@
     <h1 class="govuk-heading-xl">
       Sign in with DfE Sign In
     </h1>
-    <%= link_to "Sign in", dfe_sign_in_path, class: "govuk-button" %>
+    <%= link_to "Sign in", admin_dfe_sign_in_path, class: "govuk-button" %>
   </div>
 </div>
