+++ app/views/admin/page/index.html.erb
