+++ app/views/layouts/application.html.erb
@@ -60,7 +60,7 @@
           <nav>
             <ul id="navigation" class="govuk-header__navigation " aria-label="Top Level Navigation">
               <li class="govuk-header__navigation-item">
-                <%= link_to "Sign out", sign_out_path, method: :delete, class: "govuk-header__link" %>
+                <%= link_to "Sign out", admin_sign_out_path, method: :delete, class: "govuk-header__link" %>
               </li>
             </ul>
           </nav>
